LayoutTests/inspector-protocol/input/dispatchKeyEvent-expected.txt

-----Event-----
type: keydown
keyCode: 65
keyIdentifier: U+0041

-----Event-----
type: keypress
shiftKey
keyCode: 65
charCode: 65

-----Event-----
type: keyup
keyCode: 65
keyIdentifier: U+0041
